[Silver IV] Title: Дайн, Time: 20 ms, Memory: 17984 KB -BaekjoonHub

+++ 백준/Silver/16086. Дайн/Дайн.rs
@@ -6,18 +6,18 @@ fn main() {
     let mut tokens = stdin.split_whitespace();
     let mut next = || tokens.next().unwrap();
 
-    let n: usize = next().parse().unwrap();
+    let n: i32 = next().parse().unwrap();
     let nums: Vec<i32> = (0..n).map(|_| next().parse().unwrap()).collect();
 
     let mut cant = HashSet::new();
 
     for i in 0..n {
-        let diff = i as i32 - (nums[i] - 1);
-        cant.insert((n as i32 + diff) % n as i32);
+        let diff = i - (nums[i as usize] - 1);
+        cant.insert((n + diff) % n);
     }
 
     for i in 0..n {
-        if !cant.contains(&(i as i32)) {
+        if !cant.contains(&i) {
             print!("{i}");
             return;
         }
